+++ api/src/serge/models/__init__.py
@@ -1,0 +1,1 @@
+from .chat import Chat, Question, ChatParameters
