+++ api/src/serge/models/chat.py
