+++ api/src/serge/routers/model.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-chat_router = APIRouter(
+model_router = APIRouter(
     prefix="/model",
     tags=["model"],
 )
