+++ api/src/serge/utils/initiate_database.py
@@ -4,7 +4,7 @@ from beanie import init_beanie, Document
 from motor.motor_asyncio import AsyncIOMotorClient
 from pydantic import BaseSettings
 
-from models import Question, Chat, ChatParameters
+from serge.models.chat import Question, Chat, ChatParameters
 
 
 class Settings(BaseSettings):
